pyproject.toml
[project]
name = "grip-master-ocr"
version = "0.1.0"
description = "Real-time Screen OCR for Windows with CUDA acceleration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "pillow>=10.2.0",
    "easyocr>=1.7.1",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "pyyaml>=6.0.1",
    "pywin32>=306",
    "transformers>=5.0.0"
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }